feat: update lp dashboard and multisigner app
